Made to work via the command line. Eclipse sucks.
The entire code was riddled with errors and bugs it was changed so much. Visited the TA and asked about the issues with the command line problems. Now I am trying to download Make from GNU and test the program. I am running into a runtime error saying that the program cannot parse the latest java environment. I dont understand why!